Web層の細かいWindows依存を除去(GetModuleFileNameA/wsprintfA/型限定windows.h)
GrpImageStore.cpp の実行ファイルパス取得を GetModuleFileNameA から
SboPlatform::GetExeDirectory() へ置き換え。SessionStore.h の DWORD、
MapPartsHandler.h/SpriteSheetHandler.h の型のためだけの windows.h include を
除去(HMODULE は StdAfx.h 経由で解決される前提)。AdminWsHub.cpp/
WebSocketBridge.cpp のデバッグログ組み立てを wsprintfA から snprintf へ。

通るようになった GrpImageStore.cpp/AdminWsHub.cpp/WebSocketBridge.cpp と
単体includeが通る SessionStore.h を tools/test-sbosvr-portability.ps1 に登録。

+++ SboSvr/src/Web/AuthProvider.cpp
@@ -93,7 +93,7 @@ AuthStatus Authenticate(const HttpRequest &request, CMgrData *pMgrData, AuthCont
         }
 
         pAccountLib->Enter();
-        PCInfoAccount pAccount = pAccountLib->GetPtr(sessionInfo.dwAccountID);
+        PCInfoAccount pAccount = pAccountLib->GetPtr(static_cast<DWORD>(sessionInfo.dwAccountID));
         if (pAccount == NULL) {
                 pAccountLib->Leave();
                 return AuthStatusAccountNotFound;
